Start ChromaDB Docker Container

This function uses Docker to start a ChromaDB server container in the background.

Usage

chroma_docker_run(
  port = 8000,
  volume_host_dir = "./chroma",
  is_persistent = TRUE,
  anonymized_telemetry = FALSE,
  version = "0.6.3",
  container_name = "chromadb"
)

Arguments

port

The port for the ChromaDB container.

volume_host_dir

A string specifying the host directory to persist data.

is_persistent

Logical; whether to enable persistence. Defaults to TRUE.

anonymized_telemetry

Logical; whether to enable anonymous telemetry. Defaults to FALSE.

version

A string specifying the version of the ChromDB Docker image. Default is "0.6.3".

container_name

A string specifying the name for the Docker container. Default is "chromadb".

Value

Invisibly returns TRUE if the container is already running or started successfully.